# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
...
  errs.AppError:
    properties:
      code:
        description: |-
          Code 为唯一错误码，由 "DOMAIN.SNAKE_CASE" 格式的 ASCII 字符串构成。
          是前端做国际化映射（i18n key）和告警分组的依据。
        type: string
      details:
        description: |-
          Details 携带字段级校验错误明细，通常用于请求体（DTO）参数校验失败场景。
          如果无字段级错误，该 slice 应为 nil（JSON 序列化时被 omitempty 省略）。
        items:
          $ref: '#/definitions/errs.FieldDetail'
        type: array
      message:
        description: |-
          Message 是面向终端用户的中文提示文案，可直接渲染到 UI。
          要求：简洁、可理解、无敏感堆栈信息。前端根据此字段作为 fallback 文案展示。
        type: string
    type: object
  errs.FieldDetail:
    properties:
      field:
        description: |-
          Field 为请求参数名或 JSON path（如 "title"、"assignee.id"），
          前端可根据此字段将错误文案定位到对应表单控件。
        type: string
      reason:
        description: |-
          Reason 为该字段的具体错误原因（中文），如 "不能超过 255 个字符"、"格式不符合邮箱规范"。
          要求：简洁、无敏感信息、可直接渲染。
        type: string
    type: object
...
  knowledge.KnowledgePage:
    properties:
      content_html:
        type: string
      content_md:
        type: string
      created_at:
        type: string
      created_by:
        type: integer
      deleted:
        type: string
      depth:
        type: integer
      id:
        type: integer
      is_featured:
        type: boolean
      is_pinned:
        type: boolean
      lft:
        type: integer
      parent_id:
        type: integer
      path:
        type: string
      rgt:
        type: integer
      sort_order:
        type: integer
      space_id:
        type: integer
      status:
        $ref: '#/definitions/knowledge.PageStatus'
      title:
        type: string
      updated_at:
        type: string
      updated_by:
        type: integer
      version:
        type: integer
      view_count:
        type: integer
      workspace_id:
        type: integer
    type: object
  knowledge.KnowledgePageRelation:
    properties:
      created_at:
        type: string
      id:
        type: integer
      issue_id:
        type: integer
      page_id:
        type: integer
      relation_type:
        $ref: '#/definitions/knowledge.PageRelationType'
    type: object
  knowledge.KnowledgeSpace:
    properties:
      cover_image:
        type: string
      created_at:
        type: string
      default_permission:
        $ref: '#/definitions/knowledge.SpacePermission'
      deleted:
        type: string
      description:
        type: string
      id:
        type: integer
      is_private:
        type: boolean
      name:
        type: string
      owner_id:
        type: integer
      project_id:
        type: integer
      slug:
        type: string
      updated_at:
        type: string
      workspace_id:
        type: integer
    type: object
  knowledge.PageRelationType:
    enum:
    - referenced
    - referencing
    type: string
    x-enum-varnames:
    - RelationReferenced
    - RelationReferencing
  knowledge.PageStatus:
    enum:
    - draft
    - published
    - archived
    type: string
    x-enum-varnames:
    - PageStatusDraft
    - PageStatusPublished
    - PageStatusArchived
  knowledge.SpacePermission:
    enum:
    - viewer
    - editor
    - admin
    - owner
    type: string
    x-enum-varnames:
    - PermissionViewer
    - PermissionEditor
    - PermissionAdmin
    - PermissionOwner
...
host: localhost:8080
info:
  contact: {}
  description: 项目管理平台 Ydsz Plane 的 REST API（需求/任务/缺陷/迭代/版本/自动化/Webhook/知识库/仪表盘/全局搜索/通知等）。
  title: Ydsz Plane API
  version: 1.0.0
paths:
...
  /knowledge/search:
    get:
      description: 跨知识库空间全文检索（PostgreSQL tsvector）
      parameters:
      - description: 关键词
        in: query
        name: q
        required: true
        type: string
      - description: 空间 ID（可选）
        in: query
        name: space_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 知识库检索
      tags:
      - knowledge
...
  /pages/{page_id}/versions:
    get:
      parameters:
      - description: 页面 ID
        in: path
        name: page_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 页面版本列表
      tags:
      - pages
...
  /search:
    get:
      description: |-
        跨对象（issues/sprints/versions）全文检索 + 高亮 + JQL 结构化过滤
        JQL 示例: project:YD type:defect priority:high assignee:me()
      parameters:
      - description: 搜索词（支持类 JQL 语法）
        in: query
        name: q
        required: true
        type: string
      - description: 过滤类型 (issue|sprint|version)
        in: query
        name: doc_type
        type: string
      - description: 工作项类型 (epic|requirement|task|defect)
        in: query
        name: type
        type: string
      - description: 优先级
        in: query
        name: priority
        type: string
      - description: 状态 ID
        in: query
        name: state_id
        type: integer
      - description: 每页数量 (default 20, max 50)
        in: query
        name: limit
        type: integer
      - description: 偏移量
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 全文搜索（支持类 JQL 语法）
      tags:
      - search
...
  /spaces:
    get:
      description: 按工作空间 + 可选项目过滤，支持关键词搜索和分页
      parameters:
      - default: 50
        description: 每页数 (1-100)
        in: query
        name: limit
        type: integer
      - default: 0
        description: 偏移
        in: query
        name: offset
        type: integer
      - description: 关键词
        in: query
        name: keyword
        type: string
      - description: 项目 ID
        in: query
        name: project_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 列出知识库空间
      tags:
      - knowledge
    post:
      consumes:
      - application/json
      description: 在工作空间下新建一个知识库空间
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/knowledge.KnowledgeSpace'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 创建知识库空间
      tags:
      - knowledge
  /spaces/{sid}:
    delete:
      description: 软删除指定的知识库空间
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 删除知识库空间
      tags:
      - knowledge
    get:
      description: 返回指定知识库空间的完整信息
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/knowledge.KnowledgeSpace'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/errs.AppError'
      summary: 获取知识库空间详情
      tags:
      - knowledge
    patch:
      consumes:
      - application/json
      description: 部分更新知识库空间信息
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/knowledge.KnowledgeSpace'
      summary: 更新知识库空间
      tags:
      - knowledge
  /spaces/{sid}/pages:
    get:
      description: 获取知识库空间下全部文档的树形结构
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 文档树
      tags:
      - knowledge
    post:
      consumes:
      - application/json
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/knowledge.KnowledgePage'
      summary: 创建文档
      tags:
      - knowledge
  /spaces/{sid}/pages/{pid}:
    delete:
      description: 软删除指定的知识库文档
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 删除文档
      tags:
      - knowledge
    get:
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/knowledge.KnowledgePage'
      summary: 获取文档
      tags:
      - knowledge
    patch:
      consumes:
      - application/json
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/knowledge.KnowledgePage'
      summary: 更新文档
      tags:
      - knowledge
  /spaces/{sid}/pages/{pid}/relations:
    get:
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 列出文档关联
      tags:
      - knowledge
    post:
      consumes:
      - application/json
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/knowledge.KnowledgePageRelation'
      summary: 添加文档关联
      tags:
      - knowledge
  /spaces/{sid}/pages/{pid}/relations/{rid}:
    delete:
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      - description: 关联 ID
        in: path
        name: rid
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: 移除文档关联
      tags:
      - knowledge
  /spaces/{sid}/pages/{pid}/revert:
    post:
      consumes:
      - application/json
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/knowledge.KnowledgePage'
      summary: 回滚版本
      tags:
      - knowledge
  /spaces/{sid}/pages/{pid}/versions:
    get:
      parameters:
      - description: 空间 ID
        in: path
        name: sid
        required: true
        type: integer
      - description: 文档 ID
        in: path
        name: pid
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: 版本快照列表
      tags:
      - knowledge
...
securityDefinitions:
  Bearer:
    description: '"格式: Bearer {access_token}。登录/注册接口返回 access_token 后填入。"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"